% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComSpat.plot.R
\name{ComSpat.plot}
\alias{ComSpat.plot}
\title{Within-Community Spatial Organisation Plot}
\usage{
ComSpat.plot(
  data = NULL,
  params = NULL,
  type = NULL,
  measure = NULL,
  intervals = FALSE,
  su.size = NULL,
  ymin = NULL,
  ymax = NULL,
  xmin = NULL,
  xmax = NULL,
  p.col = NULL,
  p.cex = NULL,
  cex.axis = NULL,
  xaxt = NULL,
  yaxt = NULL
)
}
\arguments{
\item{data}{Data frame or list of data frames returned from \code{ComSpat}.}

\item{params}{Data frame providing the secondary sampling information.}

\item{type}{Character. Supply either \code{"Grid"} or \code{"Transect"}.}

\item{measure}{Character. Supply one of \code{"CD", "NRC", "AS" or
"AS.Rel"}.}

\item{intervals}{TRUE or FALSE.}

\item{su.size}{Numeric. Surface area of the smallest sampling unit (mm sq.).}

\item{ymin}{Numeric. Y axis lower limit.}

\item{ymax}{Numeric. Y axis upper limit.}

\item{xmin}{Numeric. Minimum x axis value (i.e. lower range).}

\item{xmax}{Numeric. Maximum x axis value (i.e. upper range).}

\item{p.col}{Character. Colour used for the measure. The value must be
either single or a vector corresponding to the length of data if class
list.}

\item{p.cex}{Numeric.}

\item{cex.axis}{Numeric.}

\item{xaxt}{NULL or 'n'. To control if x-axis text is displayed.}

\item{yaxt}{NULL or 'n'. To control if y-axis text is displayed.}
}
\description{
Function \code{ComSpat.plot} makes use of core R graphics systems to
display several Information Theory metrics. The \code{ComSpat.plot} does
not calculate the actual Information Theory metrics, but it accepts results
from the \code{\link{ComSpat}} function.
}
\details{
\code{ComSpat.plot} is used to construct the initial plot object. It has
the functionality to return single or multiple outputs. When multiple
outputs are returned data must be supplied as a list; a single measure for
each of the data frames will be added to the same plot. This function makes
use of core R graphics systems.
}
\examples{

# Load the training data and paramater files
data("grid.random")
data("param.grid")

# Perform ComSpat calculations
temp.rand<-ComSpat(data=grid.random, params=param.grid[1:5,], dim_max = 64, type="Grid")

# Plot ComSpat results
ComSpat.plot(data = temp.rand, params = param.grid[1:5,], type = "Grid", measure = "NRC",
 su.size = 0.01, ymin = 0, ymax = 65, p.col = "red")

# Hint - several measures can be combined using par() commands
}
\seealso{
\code{\link{ComSpat}}, \code{\link{data}}
}
\author{
James L. Tsakalos
}
