% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comspat.R
\encoding{UTF-8}
\name{comspat}
\alias{comspat}
\title{Within-Community Spatial Organization}
\usage{
comspat(
  data = NULL,
  params = NULL,
  dim_max = NULL,
  type = NULL,
  measures = NULL,
  randomization_type = NULL,
  iterations = 999,
  alpha = NULL
)
}
\arguments{
\item{data}{A matrix or data frame capturing the spatial coordinate(s) of
species sampled from a grid or transect. Each row captures the coordinate(s)
of a `Species`. If the `data` was sampled as a Transect only the `X`
coordinate is required.  If the `data` was sampled as a Grid both `X` and
`Y` coordinates are required.}

\item{params}{Data frame providing the secondary sampling information.}

\item{dim_max}{Numeric. Number of sampling units in one row of a `"Grid"` or
`"Transect"`.}

\item{type}{Character. Supply either `"Grid"` or `"Transect"`.}

\item{measures}{Vector. List the measures returned by `comspat()`. The
default option returns the compositional diversity `"CD"`, number of
realized species combinations `"NRC"` and associatum `"AS"`. Relative
associatum `"AS_REL"` is returned by default when `"AS"` is called.}

\item{randomization_type}{Character. Supply either `"CSR"` or `"RS"`.
Activating randomization initiates parallel computing.}

\item{iterations}{Numeric. Number of randomizations. The default is 999.}

\item{alpha}{Numeric. If `NULL`, p value returned. Else 1 or 0.}
}
\value{
The function returns an object of class list returning named data
frames. The variables populating the data frames are specified by the
`measures` and/or the `randomization_type` arguments. To understand the
different JNP functions we strongly recommend reviewing the work of
Bartha et al. (1998).

The following components are included within the returned object if no
`randomization_type` is specified:
\itemize{
\item{CD}{ \code{-} A matrix showing the Compositional Diversity (CD)
calculated for each spatial scale. CD measures the entropy of species
combinations. Each spatial scale is referred to as "step" and is labeled as
columns. Rows are labeled by the species involved in the calculation of CD.}
\item{NRC}{ \code{-} A matrix showing the Number of Realized Combinations
(NRC) calculated for each spatial scale. NRC measures the number of species
combinations. Each spatial scale is referred to as "step" and is labeled as
columns. Rows are labeled by the species involved in the calculation of
NRC.}
\item{AS}{ \code{-} A matrix showing the overall association
(associatum [AS]) for the collection of species calculated for each spatial
scale. AS reflects the spatial similarity and dissimilarity structure of the
grid or transect. Rows are labeled according to the species involved in the
calculation of AS.}
\item{AS_REL}{ \code{-} A matrix showing the relative association (AS_REL)
for the collection of species calculated for each spatial scale. AS_REL
reflects the spatial similarity and dissimilarity structure of the grid or
transect divided by CD. AS_REL should be used when comparing grids or
transects containing different species richness.}
\item{S_RICH}{ \code{-} A matrix showing the number of species for each
spatial scale.}
\item{H}{ \code{-} A matrix showing the Shannon diversity of species for
each spatial scale.}
}

The following components are included within the returned object if a
\code{randomization_type} is specified:
\itemize{
\item{Raw data}{ \code{-} Contains the maximum values for each of the
measures (i.e., CD, NRC, AS, AS_REL, descriptions above) for each spatial
scale obtained by randomization. The result of each randomization is
provided as rows.}
\item{Summary statistics}{ \code{-} Provides a summary for each of the
measures (i.e., CD, NRC, AS, AS_REL, descriptions above) from the
"Raw data". Summary statistics include: original value, mean, maximum,
minimum, standard deviation, coefficient of variation, p-values and
confidence intervals (all labeled as rows).}
}
}
\description{
The `comspat()` function calculates Juhász-Nagy Information Theory models.
}
\details{
The `comspat()` function presents four measures from a family of Information
Theory models developed by Juhász-Nagy (1967, 1976, 1984a, 1984b). The
measures represent co-existence relationships in multispecies communities.
For additional information on the measures please see the package vignette.
}
\examples{

data("grid_random", package = "comspat") #input data frame
data("param_grid", package = "comspat") #input parameter data frame
temp <- comspat(
  data = grid_random,
  params = param_grid[1:2, ],
  dim_max = 64,
  type = "Grid"
)
}
\references{
Bartha, S, Cz&aacute;r&aacute;n, T & Podani, J. (1998).
Exploring plant community dynamics in abstract coenostate spaces.
Abstr. Bot. 22, 49&ndash;66.

Juhász-Nagy, P. (1967). On some 'characteristic area' of plant
community stands. Proc. Colloq. Inf. Theor. 269&ndash;282.

Juhász-Nagy, P. (1976). Spatial dependence of plant populations. Part 1.
Equivalence analysis (an outline for a new model). Acta Bot. Acad Sci.
Hung. 22: 61&ndash;78.

Juhász-Nagy, P. (1984a). Notes on diversity. Part, I. Introduction. Abstr.
Bot. 8: 43&ndash;55.

Juhász-Nagy, P. (1984b). Spatial dependence of plant populations. Part 2. A
family of new models. Acta Bot. Acad Sci. Hung. 30: 363&ndash;402.

Tsakalos, J.L. et al. (2022). comspat: An R package to analyze
within-community spatial organization using species combinations. Ecography.
\doi{10.1111/ecog.06216}
}
\seealso{
The display options of [`comspat_plot()`].
}
\author{
James L. Tsakalos
}
