% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_patchy_n_isc}
\alias{grid_patchy_n_isc}
\title{Patchy no ISC grid data}
\format{
A data frame with 1551 species observations on the following three
variables.
\describe{
   \item{X}{X coordinates}
   \item{Y}{Y coordinates}
   \item{Species}{Species}}
}
\usage{
grid_patchy_n_isc
}
\description{
The package contains three grid data sets. The grid data sets
(64 × 64 sampling units) feature one simulation (grid_patchy_associated)
and two simulation derivatives (grid_random and grid_n_ISC) have been
produced by randomization. A spatially explicit individual-based simulation
model (PATPRO) was selected to produce a realistic grid community pattern
(\code{grid_patchy}) and features interspecific and intraspecific spatial
dependence due to simulated abiotic and biotic interactions. The
(\code{grid_random}) and (\code{grid_patchy_n_isc}) data sets were produced
by CSR and RS randomizations of the original simulation (\code{grid_patchy})
. The (\code{grid_patchy_n_isc}) data set only features intraspecific
spatial dependence; interspecific relationships were removed by
randomizations. The (\code{grid_random}) data set only features random
distributions; both intra- and interspecific relationships were removed by
randomizations.
}
\details{
The \code{X} and \code{Y} coordinates refer to the sampling unit of
the species in the sampling universe.
}
\examples{
data(grid_patchy_n_isc)
str(grid_patchy_n_isc)
barplot(table(grid_patchy_n_isc$Species),ylab = "Count (no.)",
space=0.5);box()

}
\keyword{datasets}
