% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param_grid}
\alias{param_grid}
\title{Parameter data for the secondary sampling of a grid}
\format{
A data frame with 16 observations (sampling steps) on the following
3 variables.
\describe{
    \item{Steps.of.scaling}{Numeric. Sampling step number}
    \item{Length.of.plots}{Numeric. Length of the sampling unit.}
    \item{Height.of.plots}{Numeric. Height of the sampling unit.}}
}
\usage{
param_grid
}
\description{
A data frame containing information on the size and shape of secondary
sampling units for the \code{comspat} function.
}
\details{
The \code{comspat} function requires spatially explicit information. The
\code{params} argument stores and provides spatially explicit information.
The parameter data frame for the grid and transect data (i.e.,
\code{params.grid} and \code{params.tran}) are available in the package.
The data contains information on the number of scaling steps
(\code{Steps.of.scaling} column) and the size (the number of primary
sampling units to create the length and height of the secondary sampling
unit, captured in the \code{Length.of.plots} and \code{Height.of. plots}
columns) of the secondary sampling units. For example, to perform secondary
sampling on a grid, the first row of the parameter data set will read 1 for
\code{Steps.of.scaling}, \code{Length.of.plots} and \code{Height.of.plots}.
If the data was sampled as a \code{"Transect"} only the
\code{Length.of.plots} coordinate is required.  If the data was sampled as
a \code{"Grid"} both \code{Length.of.plots} and \code{Height.of.plots} are
required.

\code{Length.of.plots} and \code{Height.of.plots} are measured in the
number of sampling units.
}
\examples{
data(param_grid)
str(param_grid)
}
\keyword{datasets}
