% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comspat.R
\name{comspat}
\alias{comspat}
\title{Within-Community Spatial Organization}
\usage{
comspat(
  data = NULL,
  params = NULL,
  dim_max = NULL,
  type = NULL,
  measures = NULL,
  randomization_type = NULL,
  iterations = 999,
  alpha = NULL
)
}
\arguments{
\item{data}{A matrix or data frame capturing the spatial coordinate(s) of
species sampled from a grid or transect. Each row captures the
coordinate(s) of a \code{Species}. If the \code{data} was sampled as a
Transect only the \code{X} coordinate is required.  If the \code{data} was
sampled as a Grid both \code{X} and \code{Y} coordinates are required.}

\item{params}{Data frame providing the secondary sampling information.}

\item{dim_max}{Numeric. Number of sampling units in one row of a
\code{"Grid"} or \code{"Transect"}.}

\item{type}{Character. Supply either \code{"Grid"} or \code{"Transect"}.}

\item{measures}{Vector. List the measures returned by \code{comspat()}. The
default option returns the compositional diversity (\code{"CD"}), number of
realized species combinations (\code{"NRC"}) and associatum (\code{"AS"}).
Relative associatum (\code{"as_rel"}) is returned by default when
\code{"AS"} is called.}

\item{randomization_type}{Character. Supply either \code{"CSR"} or
\code{"RS"}. Activating randomization initiates parallel computing.}

\item{iterations}{Numeric. Number of randomizations. The default is 999.}

\item{alpha}{Numeric. If (\code{NULL}), p value returned. Else 1 or 0.}
}
\value{
The function returns an object of class list returning named data
frames specified by the \code{measures} argument. Each data frame contains
species as rows and the steps of scaling as columns.
}
\description{
The \code{comspat} function calculates Juhász-Nagy Information Theory
 models.
}
\details{
The \code{comspat} function presents four measures from a family of
Information Theory models developed by Juhász-Nagy (1967, 1976, 1984a,
1984b). The measures represent co-existence relationships in multispecies
communities. For additional information on the measures please see the
package vignette.
}
\examples{

data("grid_random") #input data frame
data("param_grid") #input parameter data frame
temp <- comspat(data = grid_random, params = param_grid[1:5, ],
 dim_max = 64, type = "Grid")
}
\references{
Juhász-Nagy, P. (1967). On some 'characteristic area' of plant
community stands. Proc. Colloq. Inf. Theor. 269-282.

Juhász-Nagy, P. (1976). Spatial dependence of plant populations. Part 1.
Equivalence analysis (an outline for a new model). Acta Bot. Acad Sci.
Hung. 22: 61-78.

Juhász-Nagy, P. (1984a). Notes on diversity. Part, I. Introduction. Abstr.
Bot. 8: 43-55.

Juhász-Nagy, P. (1984b). Spatial dependence of plant populations. Part 2. A
family of new models. Acta Bot. Acad Sci. Hung. 30: 363-402.

Tsakalos, J.L. et al. (2022). comspat: An R package to analyze
within-community spatial organization using species combinations. Ecography.
doi: 10.1111/ecog.06216.
}
\seealso{
\code{\link{comspat_plot}}
}
\author{
James L. Tsakalos
}
