% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_comb.R
\name{freq_comb}
\alias{freq_comb}
\title{Frequency of species combinations}
\usage{
freq_comb(data, type, dim_max)
}
\arguments{
\item{data}{data.frame with the frequency of species combinations.}

\item{type}{Character. Supply either `"Grid"` or `"Transect"`.}

\item{dim_max}{Numeric. Number of sampling units in one row of a `"Grid"` or
`"Transect"`.}
}
\value{
The function returns an object of class data.frame.
}
\description{
The `freq_comb()` function extracts the combinational frequency of species.
`freq_comb()` is minor and features in the vignette to assist users
visualize their data.
}
\examples{

# Training data (see package vignette S.9)
s9 <- data.frame(
  Species = c("A", "D", "B", "B", "E", "C", "B", "A"),
  Y = c(1, 1, 2, 2, 2, 3, 3, 3),
  X = c(2, 3, 1, 2, 2, 1, 1, 3)
)

# create data.frame counting the species combinations
freq_comb(
  data = s9,
  type = "Grid",
  dim_max = 3
)

}
\seealso{
[`comspat()`]
}
\author{
James L. Tsakalos
}
