% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{param_grid}
\alias{param_grid}
\title{Parameter data for the secondary sampling of a grid}
\format{
A data frame with 16 observations (sampling steps) on the following
3 variables.
\describe{
    \item{Steps.of.scaling}{Numeric. Sampling step number}
    \item{Length.of.plots}{Numeric. Length of the sampling unit.}
    \item{Height.of.plots}{Numeric. Height of the sampling unit.}}
}
\usage{
param_grid
}
\description{
A data frame containing information on the size and shape of secondary
sampling units for the `comspat()` function.
}
\details{
The `comspat()` function requires spatially explicit information. The
`params` argument stores and provides spatially explicit information. The
parameter data frame for the grid and transect data (i.e., `params.grid` and
`params.tran`) are available in the package. The data contains information
on the number of scaling steps (`Steps.of.scaling` column) and the size (the
number of primary sampling units to create the length and height of the
secondary sampling unit, captured in the `Length.of.plots` and
`Height.of.plots` columns) of the secondary sampling units. For example, to
perform secondary sampling on a grid, the first row of the parameter data
set will read 1 for `Steps.of.scaling`, `Length.of.plots` and
`Height.of.plots`. If the data was sampled as a `"Transect"` only the
`Length.of.plots` coordinate is required.  If the data was sampled as a
`"Grid"` both `Length.of.plots` and `Height.of.plots` are required.

`Length.of.plots` and `Height.of.plots` are measured in the
number of sampling units.
}
\examples{
data("param_grid", package = "comspat")
str(param_grid)
}
\keyword{datasets}
